Hw web part 2

+++ src/main/java/com/internet/shop/model/Product.java
@@ -32,8 +32,8 @@ public class Product {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
-        this.price = price;
+    public void setPrice(int price) {
+        this.price = BigDecimal.valueOf(price);
     }
 
     @Override
